+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## XX.XX.XX
 * Mitigated an issue where visibility could have been wrongly assigned if a view was closed while going to background. (Experimental!)
+* Fixed a bug where passing the global content callback was not possible.
 
 ## 24.7.5
 * ! Minor breaking change ! All active views will now automatically stop when consent for "views" is revoked.